# backend/Api/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/exposition/new": {
      "post": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_New",
        "requestBody": {
          "x-name": "properties",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpositionProperties"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpositionProperties"
                }
              }
            }
          }
        }
      }
    },
...
    "/exposition/properties/{id}": {
      "get": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_GetProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpositionProperties"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_SetProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "properties",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpositionProperties"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/exposition/graph/{id}": {
      "get": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_GetGraph",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_SetGraph",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "graph",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/exposition/overlay/{id}/{exhibit_id}": {
      "post": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_SetPackageOverlay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "pkg_overlay",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageOverlayProperties"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_GetPackageOverlay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageOverlayProperties"
                }
              }
            }
          }
        }
      }
    },
    "/exposition/overlay/{id}/{package_id}": {
      "delete": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_DeletePackageOverlay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/exposition/overlay/{id}/all": {
      "delete": {
        "tags": [
          "Exposition"
        ],
        "operationId": "Exposition_DeletePackageOverlays",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExpositionProperties": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "metadata": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ExpositionMeta"
            }
          }
        }
      },
      "ExpositionMeta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PackageOverlayProperties": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "packageId": {
            "type": "integer",
            "format": "int32"
          },
          "overwriteSettings": {
            "type": "boolean"
          },
          "settings": {
            "type": "string",
            "nullable": true
          },
          "overwriteInputs": {
            "type": "boolean"
          },
          "inputs": {
            "type": "string",
            "nullable": true
          },
          "sync": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}